requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
opencv-python==4.8.1.78
ultralytics==8.0.186
insightface==0.7.3
scikit-learn==1.3.0
deep-sort-realtime==1.3.2
pandas==2.1.1
numpy==1.24.3
gunicorn==20.1.0